work on invites, socket fix

+++ modules/domain/useCase/src/main/java/com/doneit/ascend/domain/use_case/interactor/chats/ChatInteractor.kt
@@ -44,7 +44,7 @@ class ChatInteractor(
     override suspend fun updateChat(
         id: Long,
         title: String?,
-        chatMembers: List<Int>?
+        chatMembers: List<Long>?
     ): ResponseEntity<ChatEntity, List<String>> {
         return chatGateway.updateChat(id, title, chatMembers)
     }
