+++ modules/storageRemote/src/main/java/com/doneit/ascend/source/storage/remote/api/ChatApi.kt
@@ -29,7 +29,7 @@ interface ChatApi {
     fun updateChatAsync(
         @Path("id") id: Long,
         @Query("title") title: String?,
-        @Query("chat_members") chatMembers: List<Int>?
+        @Query("chat_members") chatMembers: List<Long>?
     ): Deferred<Response<ChatResponse>>
 
     @POST("chats/{id}/message")
